Use g++ to compile socket channel and worker.
This commit also splits the vsock and TCP channels.
